Moved NFT Gallery to own component
