+++ src/scenes/UserGalleryPage/UserGalleryCollection.tsx
@@ -7,7 +7,6 @@ import {
 import styled from 'styled-components';
 import unescape from 'lodash.unescape';
 import colors from 'components/core/colors';
-import NftPreview from 'components/NftPreview/NftPreview';
 import { TitleSerif, BodyRegular } from 'components/core/Text/Text';
 import Spacer from 'components/core/Spacer/Spacer';
 import breakpoints from 'components/core/breakpoints';
@@ -15,6 +14,7 @@ import { Collection } from 'types/Collection';
 import { useMemo } from 'react';
 import Markdown from 'components/core/Markdown/Markdown';
 import { DisplayLayout } from 'components/core/enums';
+import NftGallery from 'components/NftGallery/NftGallery';
 
 type Props = {
   collection: Collection;
@@ -27,17 +27,9 @@ export function isValidColumns(columns: number) {
 
 function UserGalleryCollection({ collection, mobileLayout }: Props) {
   const unescapedCollectionName = useMemo(() => unescape(collection.name), [collection.name]);
-  const unescapedCollectorsNote = useMemo(
-    () => unescape(collection.collectors_note),
-    [collection.collectors_note]
-  );
-  const columns = useMemo(() => {
-    if (collection?.layout?.columns && isValidColumns(collection.layout.columns)) {
-      return collection.layout.columns;
-    }
-
-    return DEFAULT_COLUMNS;
-  }, [collection.layout]);
+  const unescapedCollectorsNote = useMemo(() => unescape(collection.collectors_note), [
+    collection.collectors_note,
+  ]);
 
   return (
     <StyledCollectionWrapper>
@@ -52,17 +44,7 @@ function UserGalleryCollection({ collection, mobileLayout }: Props) {
           </>
         )}
       </StyledCollectionHeader>
-      <StyledCollectionNfts columns={columns} mobileLayout={mobileLayout}>
-        {collection.nfts.map((nft) => (
-          <NftPreview
-            key={nft.id}
-            nft={nft}
-            collectionId={collection.id}
-            columns={columns}
-            mobileLayout={mobileLayout}
-          />
-        ))}
-      </StyledCollectionNfts>
+      <NftGallery collection={collection} mobileLayout={mobileLayout} />
     </StyledCollectionWrapper>
   );
 }
@@ -94,27 +76,6 @@ const StyledCollectionHeader = styled.div`
 const StyledCollectorsNote = styled(BodyRegular)`
   /* ensures linebreaks are reflected in UI */
   white-space: pre-line;
-`;
-
-const StyledCollectionNfts = styled.div<{ columns: number; mobileLayout: DisplayLayout }>`
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  justify-content: ${({ columns }) => (columns === 1 ? 'center' : 'initial')};
-
-  // Can't use these for now due to lack of Safari support
-  // column-gap: px;
-  // row-gap: px;
-  margin-left: ${({ mobileLayout }) =>
-    mobileLayout === DisplayLayout.GRID ? `-${LAYOUT_GAP_BREAKPOINTS.mobileSmall / 2}px` : '0px'};
-
-  @media only screen and ${breakpoints.mobileLarge} {
-    margin-left: -${LAYOUT_GAP_BREAKPOINTS.mobileLarge / 2}px;
-  }
-
-  @media only screen and ${breakpoints.desktop} {
-    margin-left: -${LAYOUT_GAP_BREAKPOINTS.desktop / 2}px;
-  }
 `;
 
 export default UserGalleryCollection;
